refactor: extract Health Connect mappings from schema.ts
Move all Health Connect record type mappings (inbound and outbound)
into a dedicated schema-health-connect.ts file. Re-export from
schema.ts for backwards compatibility.

Reduces schema.ts from 988 to ~910 lines and isolates the HC-specific
data type mappings into their own module.
